Feature 186 refactor index (#189)
* Initial commit

* Refactoring the main browser window out of the index.js

* updating package version files

* Adding small readme for the mainAppWindow folder
