+++ .eslintrc.json
@@ -26,6 +26,7 @@
         "semi": [
             "error",
             "always"
-        ]
+        ],
+        "no-console": "off"
     }
 }
