+++ app/customCSS/README.md
@@ -1,0 +1,15 @@
+# Custom CSS
+
+This folder contains all the handlers and helpers to be able to provide custom css to teams-for-linux
+
+This is applied using config options.
+
+You can currently either use customCSSLocation or customCSSName.
+
+The customCSSLocation will load the css from that location. This should be a full path.
+
+The current options for customCSSName are "compactDark", "compactLight", "tweaks", "condensedDark" and "condensedLight". Those have been pick up from [https://userstyles.org](https://userstyles.org)
+
+Check [#77](https://github.com/IsmaelMartinez/teams-for-linux/issues/77) for more information about it and how to integrate further styles.
+
+Further information about config options can be found in the [config README.md file](../config/README.md).
