+++ app/mainAppWindow/README.md
@@ -1,0 +1,3 @@
+# Main app window
+
+This folder contains the code to handle the main app browserWindow.
